+++ pyviews/rendering/__init__.py
@@ -1,5 +1,6 @@
 """Rendering"""
 
 from .common import RenderingError, RenderingContext, get_child_context
+from .common import use_context, get_rendering_context, pass_rendering_context
 from .pipeline import RenderingPipeline, get_type, get_pipeline, create_instance, render
 from .setup import use_rendering