+++ pyviews/rendering/tests/common_tests.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock
 from pytest import mark, fixture
 
 from pyviews.core import InheritedDict, Node, XmlNode
-from pyviews.rendering.common import RenderingContext
+from pyviews.rendering.common import RenderingContext, use_context, get_rendering_context, pass_rendering_context
 
 
 @fixture
@@ -47,3 +47,28 @@ class RenderingContextTests:
         assert init_value is None
         assert self.context.xml_node == value
         assert self.context['xml_node'] == value
+
+
+def test_use_context():
+    """Should use passed rendering context in context scope"""
+    one, two, three = RenderingContext(), RenderingContext(), RenderingContext()
+
+    with use_context(one):
+        with use_context(two):
+            with use_context(three):
+                assert get_rendering_context() is three
+            assert get_rendering_context() is two
+        assert get_rendering_context() is one
+    assert get_rendering_context() is None
+
+
+def test_pass_rendering_context():
+    """Should pass rendering context as default argument"""
+
+    @pass_rendering_context
+    def get_current_context(rendering_context: RenderingContext = None):
+        return rendering_context
+
+    with use_context(RenderingContext()) as ctx:
+        assert get_current_context() is ctx
+    assert get_current_context() is None