+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added containers(Container, View, If, For)
 - added presenter
 - added pipeline name to display in error info
+- added getting current rendering context
 
 ## 3.0.1